FIX: Average all col_ meta information when merging
This caused troubles when doing maps of merged data. Now all col_ values are averaged for merged points.
If you have h scan and do binning to binds 0.5, 1, 1.5, the x values are binned but col_h values are averaged and could contain for example 0.643, 0.916, 1.76. Also the col_timer column is averaged and not summed. but this is I think correct behavior.

Change-Id: Ieebcf5b4609499414b4dfe8b9919d913d64e21fa
Reviewed-on: https://forge.frm2.tum.de/review/14178
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Tested-by: Georg Brandl <g.brandl@fz-juelich.de>