Added filtering option
In order to load polarized data, user should be able to load only selected rows. This can be achieved by filtering option.
Note: This is not working for NICOS data, because NICOS uses STRING columns for distinguishing for example different flipper state. For ILL data it is ok, as ILL uses integer field (called PAL) to distiguish states.

Change-Id: I4249fa93f07327cdfe7d23cdc895b4be5ca23775
Reviewed-on: https://forge.frm2.tum.de/review/18340
Reviewed-by: Petr Cermak <p.cermak@fz-juelich.de>
Tested-by: Petr Cermak <p.cermak@fz-juelich.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>